Resolve conflict after merge with master (+more)
Regarding the "+more" part:

I may not have realised recently, when I made changes to this file, -
that you were in-fact using the Bourne Shell (`sh`) _not_ the Bourne
Again Shell (`bash`), so I've undone my changes and hopefully once
again made it compatible with the correct shell. Sorry about that.

There were some bizarre quotations when you used `echo`, so I sorted
that out, and I think it was down to you either being confused about
rules and best practices with variables, or you wanted to have the user
see the contents of the variables _quoted_, for easy distinction; with
the latter in mind, I've single-quoted them, as seems commonplace.

Incorrect (albeit technically valid _here_) quotation:

  echo "this is a "$variable" which holds stuff"

Correct quotation:

  echo "this is a \"$variable\" which holds stuff"

Also correct, if preferred:

  echo "this is a '$variable' which holds stuff"

So why was the first one incorrect, yet still displayed?

It's partly down to how echo works, but mainly down to how shell works.
You've probably already discovered that things in shell should be
space-delimited, as that's usually how they are, by default, -
interpreted, but in this case, assuming the variable isn't assigned a
value with spaces or shell-interpretable characters, it's still valid.

Ordinarily, however, this may be a risky crutch to lean on, so it's
usually best to play it safe, in this regard.

Admittedly, it's not black and white, but I imagine after a while, -
you'll see what I'm getting at. ;)